Switching to 32 splits instead of 16 (increases performance by 27% in pyramid test, 5% in single pile ragdoll test)

+++ Jolt/Physics/LargeIslandSplitter.h
@@ -20,7 +20,7 @@ class ContactConstraintManager;
 class LargeIslandSplitter : public NonCopyable
 {
 private:
-	using					SplitMask = uint16;
+	using					SplitMask = uint32;
 
 public:
 	static constexpr uint	cNumSplits = sizeof(SplitMask) * 8;
